// LPS_Angular/src/app/models/Report/OdStatementModel.ts
export class OdStatementModel {
    NUMBER: number;
    MONTH: string;
    CREDIT: number;
    TURNOVER_CREDIT: number;
    DEBIT: number;
    TURNOVER_DEBIT: number;
    HIGHEST_BAL: number;
    LOWEST_BAL: number;
    AVG_BAL: number;
    ENDING_BAL: number;
    SWING: number;
    PER_OF_SWING: number;
    MONTHLY_INTEREST: number;
    RETURN_ITEM: number;
    DEP_CHECK_AMT: number;
    ACC_NAME: string;
    NUMBER_OF_MONTH: number;
    OD_RATE: number;
    COST_CENTER: number;
    COST_CENTER_NAME: string;
    ZONE_NAME: string;
    BRH_NAME: string;
    ACC_NO: number;
    AS_DATE: string;
    OD_LIMIT: number;
    AVG_CREDIT_TURNOVER: number;
    PER_OF_OD_UTIL: number;
    AVG_SWING: number;
    AVG_PER_OF_SWING: number;
    AVG_NUMBER_DEBIT: number;
    AVG_NUMBER_CREDIT: number;
    AVG_RETURN_CHECK_ITEM: number;
    AVG_RETURN_CHECK_AMT: number;
    AVG_RETURN_CHECK_AMT_ITEM: number;
    CIF_NO: string;
}

export class OdAccount {
  ACC_NO: number;
  ACC_NAME: string;
}